wsd: extend RequestVettingStation lifetime
We pass ownership of RVS to the DocBroker
to extend/preserve its lifetime, since
it is originally a member of
ClientRequestDispatcher, which happens to
be the ProtocolHandlerInterface for the
external (client) socket. When we pass
the socket to DocBroker, we need to
change the ProtocolHandlerInterface to
that of WebSocketHandler, since we
upgrade it to WebSocket. To do this,
we must ensure that RequestVettingStation
is preserved.

Change-Id: I4b0413fed2346446887b7436cc4d34791dcddfb5
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>